+++ HISTORY.md
@@ -2,6 +2,7 @@ unreleased
 ==========
 
   * Add `priority` to `cookie` options
+  * Fix handling errors from setting cookie
   * Support any type in `secret` that `crypto.createHmac` supports
   * deps: cookie@0.5.0
     - Fix `expires` option to reject invalid dates